Notify-ASCRecommendationsbyAzureActivity
Hat tip to Michael Makhlevich and Andrey Karpovsky who exposed me to this idea.

Using the Azure Subscription Activity Logs and looking for those that did a /write call within last 7 days on the resource in ASC recommendation to send the email to. This should indicate ownership or responsibility to resource and could make changes to remediate.

If possible may need some community testing against more ASC Recommendations and scenarios.