Create MDE Device Tags from Azure Defender for Servers
This Logic App can be set to run daily, weekly. Upon scheduled trigger it will match Azure VMs to MDE Devices and Set a defined MDE Device Tag on the Server in MDE. This can be useful to help with reporting in MDE portal and MDE Tag can also be tied to a Device Group so you can Separate Permissions to Servers and also set Automation Investigation & Remediation (AIRs) to none, Semi, or Full for the Servers onboarded to MDE from Defender for Servers P1/P2.